added predelay control (upto 500ms) to the DattorroReverb
- added predelay control
- fixed delayBuffer : bufferLength = number of samples + 1

added the samplePerMsec as a member of the APF strutures

- avoids 1 recalculation everytime setParameters() is called when updating parameters with plugin parameters

+++ SchroederReverb/Source/PluginProcessor.cpp
@@ -88,8 +88,8 @@ void SchroederReverbAudioProcessor::processBlock(juce::AudioBuffer<float>& buffe
     // update Reverb Algorithm parameters
     reverbAlgorithm.setParameters(reverbControl);
 
-    float rightIn = 0.0;
-    float leftIn = 0.0;
+    double rightIn = 0.0;
+    double leftIn = 0.0;
 
     for (auto sample = 0;sample < buffer.getNumSamples(); ++sample)
     {
